python0server/requirements.txt
flask
minio
requests
python-dotenv
Werkzeug
